c/controller_backend: allow leader replica to finish update
Using delta operation retry counter to chose a replica that is going to
dispatch finish update command. Dispatching the command multiple times
does not cause any harm as it is idempotent.

Previously only nodes that were added to the replica set were allowed to
dispatch finish update command. This approach is to restrictive as any
of the replicas from current assignment can do it safely.

This approach fixes situation in which partition move cancel would never
finish in situation when node added to replica set as a result of
cancel operation is offline.

Signed-off-by: Michal Maslanka <michal@redpanda.com>