config: new properties for quota balancer
Add new cluster config properties required for intra-node quota
balancer: balancer run period, and minimum shard TP quota in two
forms, ratio of node limit, and absolute value. For the ratio, add
a validator that limits double values between 0 and 1. Properties
are bound into snc_quota_manager.

Note on the ratio property.

It may be more useful than the absolute bps value, because it basically
says, e.g. "5% of the shard's equal share of the node wide quota is
always assigned to each shard no matter what". It's easier to reason
about than the absolute value: we can see then that 95% can be taken away,
which means that in an extreme case one shard can get up to 95%
of the node wide limit.

But ultimately it's about convenience: since the strongest of the ratio
and the absolute value is applied, it's up to the user which one to use.