k/quotas: balancer timer
Balancer worker is a timer with spawn_with_gate for the actual
processing, and a semaphore to avoid reentrance into the balancer
routine. Timer is only active in one shard. Timer period is measured
between the moment of balancer worker invocation and is therefore
independent of the time it takes for the balancer to run. Minimum
balancer period is enforced.