kafka: Handle for case if node is isolated
If node isolated or decomissioned it can not handle kafka requests from
client, so in this case we need to signal client comunicate with another
broker. For this we need to exclude isolated node from brokers list and
return -1 for controller_id, after it client will send metadata request to
another broker and will comunicate with it

Also we can not put isolated node like leader for partition.
To prevent client stuck we gonna add fake leader to force
client connect to another broker