c/topic_dispatcher: adjusted p. allocator updates when move is finished
When partition is being moved new replicas are allocated on the nodes
were the partition replicas are moved to but until the move finishes the
old replicas are still present on the nodes removed from the
replica set. Updated topic dispatcher logic to reflect that. Now we
remove allocation from partition allocator only in finsh moving command
handler.

Signed-off-by: Michal Maslanka <michal@redpanda.com>