c/topics_table: handling revert cancel partition move command
Added handling for `revert_cancel_partition_move` command to the topic
table. The command is going to be emitted by controller_backend when
partition underlying raft group reconfiguration that is requested to be
canceled is already finished. The command finishes move and updates the
topic partition with replica set aligned with the target of original
move command.

Signed-off-by: Michal Maslanka <michal@redpanda.com>