c/controller_backend: emit revert cancel command
Controller backend is now dispatching replication of revert cancel
partition move command when partition reconfiguration finished before
backend had a chance to cancel it. When supported the revert command is
used by the backend instead of reconfiguring the partition raft group to
the replica set from before the requested move.

Signed-off-by: Michal Maslanka <michal@redpanda.com>