cluster/types: add local retention disabled checks
Fixes #8344

This commit adds a new method to the tristate type, 'is_engaged', which
checks if the tristate is in the "Disabled" or "Set" state. We often use
the "Not set" as the default state and it's useful to check if changes
have been applied.

The new method is then used in 'topic_properties::has_overrides'. This
fixes a latent bug where 'retention_local_target_ms' and/or
'retention_local_target_bytes' being disabled was ignored by
'topic_properties::has_overrides'. The bug never manifested as we
override 'cleanup_policy_bitflags' for all Kafka topics.