cloud_storage: cleanup segment reader on shutdown
We previously only called set_end_of_stream() when met with an unexpected
exception, citing that the stream is left in an undefined state. It
seems reasonable to do this with gate_closed_exceptions, since we do
this explicitly from the loop if the gate is closed.

(cherry picked from commit d7e3f19bbbdc0afab3a02ebe9daff3494fad9bc6)