k/produce: return not leader for partition when shard is not found
When moving partition the metadata may not reflect the actual replica
placement as the actual data movement may take time. If partition
metadata is correct but the partition not yet exists on a node we must
return retryable error to the client. There is a possibility that during
handling next client request partition will already be available on the
node.

Changed error returned when partition shard wasn't found from
`UNKNOWN_TOPIC_OR_PARTITION` which is not retryable to
`NOT_LEADER_FOR_PARTITION` which forces producer to retry.

Fixes: #8290

Signed-off-by: Michal Maslanka <michal@redpanda.com>