r/consensus: do not cancel reconfiguration when nodes are about to be removed
When reconfiguring raft replica set at some point it does not make sense
to cancel the reconfiguration even if the cancellation was requested.
When configuration change is in the stage where nodes that are to be
removed are demoted to learners it makes more sense to let the
configuration change finish as canceling the change would require more
larger quorum.

Detailed description:

Consider configuration change from replica set A to replica set B.

Assumptions:

- `A + B` denotes a set union
- `A - B` denotes a set difference

 Raft configuration is presented as:
 ```
 {'c': {'v': [A], 'l': []}, 'o':{}}
 ```
Where:

`c` - current configuration
`o` - previous configuration (present when configuration is in joint
consensus)
`v` - voters
`l` - learners
`to_add` - nodes to add to configuration
`to_rm` - nodes to add to remove from configuration

Configuration change requires following steps:

```
1. {'c': {'v': [A], 'l': []}, 'o':{}}
2. {'c': {'v': [A], 'l': [B-A]}, 'o':{}, to_add: (B-A), to_rm: (A-B)}
3. {'c': {'v': [A + B], 'l': []}, 'o':{}, to_add: (B-A), to_rm: (A-B)}
4. {'c': {'v': [B], 'l': []}, 'o':{'v': [A + B], 'l': []}, to_add:
   (B-A), to_rm: (A-B)}
5. {'c': {'v': [B], 'l': []}, 'o':{'v': [B], 'l': [A-B]},to_add: (B-A),
   to_rm: (A-B)}
6. {'c': {'v': [B], 'l': []}, 'o':{}}
```

In each of the steps the group needs following nodes to be available and
continue reconfiguration:
```
1. availability: majority(A)
2. availability: majority(A), required to continue reconfiguration: (B-A)
3. availability: majority(A+B)
4. availability: majority(A+B) AND majority(B)
5. availability: majority(B)
6. availability: majority(B)
```

When canceling in different steps of reconfiguration we must consider
following scenarios:

Canceled in 2.

Change in availability

    Before:
        availability: majority(A), required to continue reconfiguration: (B-A)

    After:
        availability: majority(A)

Required steps:
```
1. {'c': {'v': [A], 'l': []}, 'o':{}, to_add: (), to_rm: ()}
```

Canceled in 3.

Change in availability

    Before:
        availability: majority(A+B)

    After:
        availability: majority(A+B)

Required steps:
```
1. {'c': {'v': [A+B], 'l': []}, 'o':{}, to_add: (), to_rm:(B-A)}
2. {'c': {'v': [A], 'l': []}, 'o':{'v': [A+B], 'l': []}, to_add:(),
   to_rm: (B-A)}
3. {'c': {'v': [A], 'l': []}, 'o':{'v': [A], 'l': [B-A]}, to_add: (),
   to_rm: (B-A)}
4. {'c': {'v': [A], 'l': []}, 'o':{}, to_add: (), to_rm: ()}
```

Canceled in 4.

Change in availability

    Before:
        availability: majority (A+B) & majority (B)

    After:
        availability: majority (A+B)

Required steps:
```
1. {'c': {'v': [A + B], 'l': []}, 'o':{}, to_add: (A-B), to_rm:(B-A)}
2. {'c': {'v': [A], 'l': []}, 'o':{'v': [A + B], 'l': []}, to_add:
   (A-B), to_rm: (B-A)}
3. {'c': {'v': [A], 'l': []}, 'o':{'v': [A], 'l': [B-A]}, to_add: (A-B),
   to_rm: (B-A)}
4. {'c': {'v': [A], 'l': []}, 'o':{'v': [A], 'l': []}, to_add: (A-B),
   to_rm: (B-A)}
5. {'c': {'v': [A], 'l': []}, 'o':{}, to_add: (A-B), to_rm: (B-A)}

```

Canceled in 5. and following

Change in availability

    Before:
        availability: majority (B)

    After:
        availability: majority (B), required to continue reconfiguration: (A-B)

```
1. {'c': {'v': [B], 'l': [A-B]}, 'o':{}, to_add: (A-B), to_rm: (B-A)}
```

Since the conditions to finish canceled reconfiguration are more strict
than for continuation of the configuration change we change the raft behavior
to prohibit canceling reconfiguration after step 4 finished.

Fixes: #8301

Signed-off-by: Michal Maslanka <michal@redpanda.com>