k/quotas: 0 balancer period enforces default per-shard quotas
Whenever balancer is disbled on the fly by changing
`kafka_quota_balancer_node_period` to 0, we would like whatever
effective shard quotas have been set by the balancer by then to be reset
to the defauld shard quotas (as if no balancer has ever been working
on them). This is an intuitive effect of trasitioning to the
configuration with balancer disabled ("balancer disabled" =>
"shard quotas are not balanced" => "shard quotas are the same").

In case a user wants to keep the effective quotas,
they have an option to specify a very long balancer period
to pretend that it works.