archival: reset probe when leadership change detected
In the move from reconciliation-based archival management to
one-per-partition, we lost the property that probes unregistered their
metrics during leadership changes.

This is particularly problematic for metrics that are updated with
values rather than deltas. Such metrics, in aggregate, could be
double-counted across the cluster.

This commit updates the probes to reset themselves upon changing terms.