storage: share target replay bytes amongst shards
'storage_target_replay_bytes' is a configuration option that represents
a soft limit of how many dirty bytes the state machines on a node can
have. It is meant to be shared between the shards.

This commit fixes a bug due to which this target was not being shared
between shards. The impact is that we were replaying more data (up to
the 64MiB hard-coded limit) from the log on restarts as the latest
snapshot was further in the past.

Fixes #7256
Related #6854