Restrict elems-per-frag to 2^n in frag vector
Restrict the number of elements in a full fragment to the highest power
of two that still satisfies the max_frag_bytes limits.

This keeps the indexing for operator[] considerably similar because
a shift is needed instead of a division-by-constant.

While we're in there, add an accessor for the number of elements held
in each fragment, allowing users to query this constexpr value.