Add test coverage to fragmented vector
Add several new tests for fragmented vector, covering:

 - Assignment from std::vector
 - clear() method
 - const vs non-const iterators
 - Fragment sizing
 - Iterator arithmetic and comparison
 - Using vector in std::sort

Furthermore, a wrapper is created which checks the consistency of the
vector on every operator->() deference: by using this widely in the test
we fail fast when an inconsistency is discovered.

We use an internals-accessing friend class to enable some of this, e.g.,
the consistency checking, without needing to make internals public.