cloud_storage: Change remote list API signature
The remote list method used to return the items found, but it is
possible that this API may be used to perform more complex operations
on a bucket, such as listing objects with a delimiter to traverse
bucket hierarchies. For this more fields of the result of listing the
bucket are required. The remote list API is changed to return
additional information collected from the API call, in addition to the
bucket contents.