cloud_storage: Make focused calls to list API
During topic recovery in order to collect possible topic manifest files,
the behavior is changed to make multiple calls to list api with
restrictions, such that first the top level items are listed, then
potential topic meta prefixes are listed, and finally the manifest
paths found in those prefixes are collected. This avoids trawling
through a large number of files to find a few topic manifests.