utils: Add timed_moving_average
This is a moving average wich takes timestamps into account. Every
update call accepts a timestamp. The width of the sliding window is
defined as a time interval. There is also a resolution parameter which
limits precision of time calculations. If the width of the sliding
widnow is 10s and the precision is 100ms the object will use 100 buckets
to store samples (10s / 100ms). This means that the time increments
which are smaller than resolution won't be taken into account and that
the memory requirements can be lowered by lowering the precision.