archival: Add 'set_enabled' and 'stop' methods
to the 'archival::housekeeping_job' interface. The 'set_enabled' method
changes the state of the job which now can be enabled or disabled. The
'stop' method is used in situation when the 'ntp_archiver' is stoppped
and need to wait until all jobs that belong to it are not in use.