schema_registry/protobuf: Early exit recursion if file exists
During a depth-first search for references, it's possible for a
reference to be looked up and added to the `DescriptorPool`
multiple times, which fails.

If the `DescriptorPool` contains the reference already, early exit
without a looking up the referenced schema, and adding it to the dp.

Signed-off-by: Ben Pope <ben@redpanda.com>
(cherry picked from commit 62caac33f55a95642304db70e8365bfde49f16a4)