kafka: wait for offset retention feature
In the upgrade case it is nice to wait for all nodes to have offset
retention feature available. The reason is because offset retention
policy is applied independently by each node. In a mixed version cluster
it's likely that nothing crtically bad would happen, but what is more
likely, is that if the mixed version situation persisted, one could
imagine anomalous behavior with some groups experiencing retention and
others not or worst case some offsets reappearing.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>