c/members_manager:do not recommission a node when it is demoted to learner
A recent change in raft configuration cancellation logic prevents raft
re-configuration canceling when nodes to be removed are demoted to
learners. Since 'moving back' is not longer possible we should prevent
recommissioning a node when it is already a learner in previous part of
raft joint configuration.

The fix doesn't fix the race condition entirely but it reduces the
possibility of it occurring. The race condition will be fixed with an
introduction of simplified raft configuration. Also the race condition
isn't really breaking anything. A user may receive success from
recommission API however the node will simply be removed. It is a
behavior that we accepted for partition movement cancel API.

Fixes: #8179

Signed-off-by: Michal Maslanka <michal@redpanda.com>