storage: track if batches have monotonic stamps
This commit extends the on disk segment index with a new field that
tracks whether the last timestamps of batches within a segment are
monotonically increasing. This information is required in order to
determine if the relative time index can be used for time querying (see
following commits).

For segments that were written by a previous version of Redpanda, we
assume that non-monotonic timestamps in order to preserve the existing
timequery semantics. For newer segments, we rely on the state in the
index.