k/quotas: ignore minimum quota while updating effective quotas
This change streamlines the policy around miminum quota setting:
it is only effective when quota balancing is done, but the node wide
limit, when it causes default shard quota to be less than the minimum,
will be set that way -- i.e. minimum quota does not override
limit settings.

This principle was implemented in some places already, this change
also makes it work when limits are decreased by the configuration.
Instead, the absolute quota limit of 1 is appiled.