rpc/rpc_server: set promise on exceptions in `dispatch_method_once`
Changes the logic of `dispatch_method_once` to avoid a broken promise
when the connection is closed before the input is skipped for unknown
methods and unknown versions. This ensures that we are always setting
the body_parse promise after reading in all the input and before we send
a reply.

(cherry picked from commit e47ab0dccb51a1a07e351219c94eb30575d42d82)