c/raft0_utils: do not throttle raft0 learners recovery
Since raft0 log is usually small and it is required to up to date on all
of the nodes to provide full functionality of the cluster it makes sense
to disable throttling of its learners recovery. This way nodes joining
cluster will be faster fully up to date.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit f79cefd3507976957fd7f26451baeb0786e10d54)