tests/docker: split out slow building deps file
When we created ducktape-deps.sh, there was an unintended
side effect that any dependency change caused a full rebuild
of the container (including the ultra-slow OMB build).

This commit splits out the dependency definitions into
a "fast" and "slow" file.  This is fairly coarse grained,
but it gets us back to a place where we can change our
verifier code without having an hour long container rebuild
for all the developers who check out the code and run a test.