k/quotas: fix uneven default quota distribution +refactoring
Refactoring/simplification: `quota_balancer_update()` does not take
the special `dispense_quota_amount` structure as a parameter, now
it takes the old and the new value for default node quota. This has
allowed to discard a good deal of complicated code.

This has also allowed moving the code that analyzes the difference
between the default quotas into `quota_balancer_update()` and to fix
an issue when the newly set quota (after it was unset) cannot always
be divided between the shards in full because of the limitation imposed
by `dispense_quota_amount`.

`tristate` is used to represent the quota value that may be either
present, absent, or does not need a change. To facilitate that usage,
`tristate`'s `get_optional()` is made public.