archival: Improve adjacent segment merging
Attempt to run up to 4 uploads per round. This will allow housekeeping
to keep up with uploads when there is a large backlog of pre existing
small segments.

Fix the bug when the segment which is large enough so it can't be merged
with the next segment prevents the segment merging from progressing.