k/quotas: guard against passing negative quotas to TB
bottomless_token_bucket would vassert if it has a negative quota
passed as an argument. Since the value of an updated quota comes
as a result of nontrivial calculations while handling cluster property
updates, this extra guard ensures that negative (or 0) quota value
never hits the token bucket; an error will be logged instead.