s/compacted_index: add parsing of v1 footers
An unfortunate side effect of introducing v2 compaction footers was that
perfectly good indices with v1 footers would have to be rebuilt (after
restart we have to verify index integrity to verify that the segment is
already compacted and indices with v1 footers failed the version check).
This commit adds compatibility code for v1 footer parsing.