features: add await_feature_then
It isn't uncommon to want to wait for the activation of a feature, and
then do something once activated (e.g. start using the feature).
I intend on using this exact pattern to introduce a new error code that
isn't compatible with older versions.

This commit encapsulates the wait logic in redpanda/application.cc to be
a first-class citizen of the feature_table.