r/replicate_entries_stm: do not skip sending request to removed follower
When new simple configuration is replicated after leaving joint
consensus we still may send a request to follower that is going to be
removed after the transition from joint to simple state. This way there
is a chance that the follower will receive an information about it being
removed from the cluster. The change introduced here is just an
optimization and it does not influence correctness or configuration
change semantics. It just allows the removed follower to discover its
removal.

The last configuration change  delivery isn't guaranteed and it may
happen that the follower will not receive it so we should not treat it
as reliable mechanism to discover removal.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 76a9c4e525ef69b04cd24ba1825b9e7f21d14f4c)