redpanda: hexdump config on 'unknown escape character'
On this particular exception type (but not on general config load
errors), dump out a hexdump-style view of the bytes we loaded
from the config yaml file.

This error occurs very occasionally, it is suspected that something
is wrong with lower layers in the system, but it's hard to
infer much without knowing what's in the byte buffer
that we're reading.

Related: https://github.com/redpanda-data/redpanda/issues/3798