Remove .reserve() call on partition metadata
reserve() was being used on the std::vector of partition metadata to
size it correctly, but now that we are using fragmented_vector this
method no longer exists. It does not make sense for fragmented_vector
since that class does not copy data to a new contiguous area as it
grows, and always immediately allocates each fragment as its maximum
size, so reserve would not change the allocation or copy pattern at all.