utils: Fix hang in input_stream_fanout
If the source stream throws an exception readers of all fanout streams
will hang indefinitely. The exception won't be processed and the
background producer fiber will exit with ignored exception. As result
the reader will wait for new data indefinitely if the 'stop' method is
not called.

This commit changes this behavior. The exception in the background
producer thread is saved and bubbled up to readers if the source stream
throws.

(cherry picked from commit ff385aeb18209524d04d0bd27df4ec6ffbd58b5b)