k/quotas: fix negative shard quotas +refactoring
The algorithm that distributes negative quota deltas between shards
is rewritten. The principal change is that it does not attempt to
take any balancer role and does not use current shard surpluses
to decide how much of the quota to take away from each shard.
Instead it relies solely on the effective shard quota, dispensing the
delta pro rata to the effective quotas and thus tries to maintain
the current quotas profile between the shards.