k/handlers: add a check for the number of replicas
Currently, Redpanda does not allow an even number replication factor. In
the case that a user attempts to dereplicate the replication factor to
an even number (e.g., 3->2) by removing a node from the replica set,
then the partition allocator will throw. However, in the case that a
user attempts to upreplicate to an even number (e.g., 3->4) by adding a
node to the replica set, the partition allocator will select 3 out of
the 4 nodes for the new assignment.

This commit adds validation that checks if the incoming request matches
the topic's replication factor. Returning an error at this level is
preferred over letting the request get to the allocator.

It is important to note that Kafka does support
upreplicating/dereplicating to an even number replication factor.
Therefore, this changes introduces a divergence from Kafka.