storage: use actual timestamp for index truncation
Previously, when performing suffix truncation, we used the closest
max timestamp in the segment index as the max timestamp for the segment.
Since, we only index a batch once every 32KB, the latest timestamp in
the index is not a great heuristic (later batches may have greater
timestamps). This is problematic as the maximum timestamp in the index
is used for enforcing time based retention.

This commit improves the heuristic by using the max timestamp of the
lates batch before the truncation point.