kafka: replace default commit time on stored offsets
When an offset is stored with a default commit time (-1) kafka replaces
the value of commit time with the current time before saving the offset
to the log. Redpanda had been storing the value unmodified. Offset
retention uses this value to calculate the offset expiration.

Prior to v23 both commit and expiration timestamps were unused, by the
kafka layer, so it is safe to make this change. The timestamps are now
used by offset retention.

Since retention time was dropped prior to v23 they are not automatically
reclaimed as a conservative measure, but can be manually reclaimed using
kafka's offset delete api.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>