tests/acls: ignore disconnect errror from GnuTls
Previously the upgrade started on RP v22.1.x but that version does not
handle errors from the GnuTls lib. This is what led to a BadLogLines
error in #7536. The fix is in PR #7212 but that is not in v22.1.x.
Therefore, this commit adds the log line to the allow list.

Fixes #7536