features: introduce new version epoch batch type
The version epoch batch type is intended to be added to partitions to
establish a barrier marking a cluster version change event. For example,
a version fence added to a partition may be used to implicity categorize
all batch preceding the barrier as being written prior to the upgrade.

Currently this batch type is added manually to partitions that require
it, and should be used to establish an upper bound position for an
upgrade event.

The type itself is reserved in the offset translator and is exempt from
compaction.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>