config: fixed min boundary for kafka_throughput_limit_node_out_bps
The minimum boundary for thoughput limit properties was originally
`ss::smp::count` so that each shard would not get a zero default quota.
Later that was changed to dynamic boundary control, while the static
property minimum was changed to a constant `1` so that there is no
situation when cluter properties replayed from the controller log
were out of property bounds if smp::count on the node is different
from the node where the configuration was made.

However the change for kafka_throughput_limit_node_out_bps was left behind.