tests: enable node TP limits
Node wide TP limits are enabled in:

ManyClientsTest: node TP limit is set at approximately the same level
as the exising per-client limit, with intent to supercede it in future,
to guard against memory overflows in redpanda when clients are on
larger instances

ManyPartitionTest: node TP limit is set at 3 times of expected_bandwidth
per node. This is done in hope that (1) this limit will not impact
the test results, and (2) the limit will be reached occassionally
and the quota balancer will be doing its job. If that does not happen,
the factor of 3 may be lowered.