node_operations_fuzzy_test: fix configs for older versions
In the 000db2c78ac861098b3dc6f1eba32613dae14716 backport, we started
using the NodeOpsExecutor in the fuzz test, meaning we no longer track
next node ID.

This updates the executor to keep track of the next node ID as we used
to, to allow configuring the old configuration style.

NOTE: this commit only lives on v22.3.x, as `dev` does not suffer from
this problem.