rpc: guarantee that resource_units don't outlive the request
An important RPC invariant is that resource_units supplied by the caller
must not outlive the request (otherwise it is possible that corresponding
semaphores are already destroyed by the time we try to return
resource_units).

Previously breaking this invariant was possible if the request timed out
while resource_units were owned by the lambda emplacing the request to
_requests_queue (this lambda can wait for the _memory semaphore units
for substantial time).

This commit makes it easier to guarantee this invariant by storing
resource_units in _correlations instead of _requests_queue. This way if
we complete the request due to timeout or disconnect, resource_units are
automatically released when we remove the entry from _correlations.

Fixes https://github.com/redpanda-data/redpanda/issues/8421
Fixes https://github.com/redpanda-data/redpanda/issues/8619