resource_mgmt: reduce IO priority of tiered storage reads
This is the priority used in cache::put when writing to
local disk while promoting segments from object storage
to the cache.

Currently, this is done at the same priority as raft
writes (i.e. Kafka Producers), and a consumer can easily
overwhelm a producer by e.g. consuming from a 20-partition
topic on a node with a cold cache: the 20 concurrently-written
cache objects dominate disk I/O.