c/controller_backend: always allow canceled operation to be executed
Executing canceled operation may be required to create an initial
replica for changed partition. After introduction of reverting cancel
operation it is not always the case that the canceled operation may be
skipped. If the cancel operation was reverted we MUST not skip execution
of update operation. Change the way how partition operations are
interrupted. Now they are interrupted after the initial update was
executed. This way when canceling an operation the partition replica
created by update operation will be instantiated.

Fixes: #8657

Signed-off-by: Michal Maslanka <michal@redpanda.com>