archival: Control lifetime of jobs from housekeeping svc
Previously, the owner of the housekeeping job was required to call
'deregister_job' method of the workflow. This worked correctly for jobs
which are not executed at the moment. The running jobs were supposed to
hold a gate which in turn is closed by the 'stop' method of the job.

The problem with this approach is the following one. If the job is
executed and the owner of the job needs to stop the owner calls
'interrupt' to stop the job early and then 'stop'. The 'stop' call waits
until the 'run' method of the job exits. After that the job could be
disposed by the owner. But the 'housekeeping_workflow' may be suspended
at the moment and when its background fiber will continue it will try to
check the state of the running job. In order to do this it will call
'interrupted' method of the job. If the job is alredy disposed the
applicaton will segfault.

To fix this the following changes were introduced. The
'housekeeping_job' interface now has the 'release' and 'acquire'
methods. The 'acquire' is called by the housekeeping workflow when the
job is registered and the 'release' is called when the job is removed
from the queue. This way the housekeeping workflow can inform the owner
of the housekeeping job that the job will no longer be used. The
workflow guarantees that after the call of 'release' the job won't be
touched by the service. The 'stop' method of the job has to block until
the 'release' method is called (if the 'acquire' was called in the first
place).

Also, the 'housekeeping_workflow' was reorganized a bit. Instead of
using one queue and the optional<reference_wrapper<housekeeping_job>> to
store currently running job the object now has three intrusive lists.
One for pending jobs, one for running job, and the last one for executed
jobs. The job is moved between the lists. When it's executed it is moved
from pending to running list. When the execution is completed it's moved
from running list to executed list. And when the next housekeeping cycle
stars it's moved from executed to pending list.