cloud_storage_clients: Handle ABS next marker
ABS list response contains a next marker which is empty if the result is
not truncated. If this field has a value then the result is truncated,
and the marker value should be used in the next list call to fetch more
results. This change adjusts the xml parser to correctly store and use
the next marker field.