rpk: avoid writing cfgFile if there are no changes
If the contents of the file in rpk redpanda start
are the same after running the command, rpk should
not write to the file.

In the past, if you have the same content but
ordered it in a different manner, rpk will rewrite
and reorder the configuration file.

This commit introduces rawFile field in the
config struct, this is a in-memory representation
of the loaded raw file.