s/segment_set: ignore config segments in timequery
This commit updates the logic that fetches the segments eligible for a
timequery to use filtered binary search in order to exclude the segments
that only contain configuration batches.

We wish to do this because said segments always have wall timestamps,
but user provided data can come with it's own timestamps specified on
the producer side. This used to trip timequery up.