k/quotas: synchronize quota reset on balancer disable
When the quota balancer is disabled, that should reset effective
quotas. In #8392, it was done asychronously to balancer updates,
which proved incorrect in case of simultaneous change of many
related properties. This change makes use of quota_balancer_update()
to deliver quota reset values to shards in synch with other updates
and the balancer.

A part of comment from the older code moved to a more appropriate place.