cloud_storage_clients: make delete op sequential
The ABS batch delete wrapper is changed to delete items in sequence.
Deleting in parallel by scheduling coroutines runs into issues with the
underlying client. There is a batch blob API but it requires changes to
azurite which does not seem to transparently support the batch
operation model.

This change makes the delete operation much slower, so eventually we
should aim to move to the batch blob model.