txn: make try_abort fail faster
When data partition expire a tx it gets a lot and send a request to
the txn coordinator. If it happens at the same time as the leadership
change the request may land on the old coordinator and it may stuck
in the retry loop until it finally expires and while it happens the
lock on the data partition will prevent the new coordinator from making
the progess. Removing the retry loop to fix it.

It's safe to remove the loop because the data partition has its own
retry loop.

Fixes https://github.com/redpanda-data/redpanda/issues/8726