tests: retry kgo remote call if connection hangs
I saw a remote call hang in sending the last_pass signal. It isn't quite
clear why this was the case (perhaps the kgo server was too busy
consuming? it didn't seem blocked on consuming, but it did fetch many
offsets), but this commit adds a timeout to the remote call, and retries
periodically, since presumably it is transient.