many_partitions_test: add ok_to_fail
The tiered storage variant of the test isn't fully stable, and has a
known failure mode. So we at least still see some coverage before
releasing this marks the test as ok_to_fail with the expectation that
the failures will be looked into shortly.

Additionally, it was found that shutdown can take a while (though it
doesn't appear to hang), so this extends the timeout for shutting down.