r/types: fixed serialization of mixed failed and successful hb replies
When there was one failed heartbeat reply in the array of heartbeat
replies it might cause all other heartbeats to fail as the not
initialized field for target node_id was used for all of the heartbeat
replies. Heartbeat reply serialization uses a varint delta encoding to
minimize the size of exchanged messages and de-duplicate information.
Changed the way how the target_node_id field is selected for
serialization. Instead of using just a first node we will use last one
after the array of responses is sorted for delta encoding. After sort
operation the largest value of `last_flushed_offset` is at the end of
the array as not initialized value is a `model::offset::min()` if there
is any other value available it will reside at the end of the array
after the sort.

Fixes: #8733

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit e89d69a106029d0e66b1746ab1922aa4cacc54cb)