s/segment_reader: gate in-flight operations
This commit introduces a gate that guards against the closing of the
segment while there are in-flight operations that cannot be pre-empted
(for instance, creating new 'segment_reader_handle').

Previously, we ran into use-after-free scenarios where the 'segment'
(and implicitly the 'segment_reader') was removed with oustanding
futures in the reader. This could happen if we were attempting to
create a 'segment_reader_handle' while the segment was destroyed.

We've seen this scenario occur during compaction either due to the
segment getting erased (adjacent compaction) or due to the reader being
swapped out.