storage: Out of order offset_translator_state
Make offset_translator_state process gaps out of order with the
condition that the out of order data is the same as previously seen by
the offsset_translator_state. This is needed to handle inconsistencies
in the manifests, e.g. overlapping segments.

Fix problem in remote_segment that may cuase the state of the reader to
go out of sync with offset translator state.