TS timequeries: bail on if segment not found
In the case a timequery is made with a timestamp later than any
segment in the TS index, we return an end iterator from seek_segment
and do not check for this condition subsequently in find_cached_reader
which causes an assert when an end() iterator is de-referenced.

This change checks for an end() iterator and bails out, treating it the
same way as other early exist conditions such as if the partition has
no associated segments. This will fall back to the local log to
satisfy the time query.

Fixes #8811.