e2e_topic_recovery_test: have s3 check the delta offset
The test currently waits for a target Redpanda offset to be uploaded,
and then checks that the correct number of Kafka offsets are consumable.
This isn't quite right, since we also need to take into account config
batches in our Redpanda offsets.

This updates the test to take into account the 'delta_offset_end' field
of the last offset, which should give us an accurate view of the number
of Kafka data batches in the topic.